Clear list after notifying list model to prevent array access error
The list model will indirectly access each index at which an item has been removed
thus clearing the list beforehand causes an error.

It should be safe to clear the list later since we are repopulating it anyways and
the list is not accessed before it is repopulated once we have notified the list model.