feat: narthex compatible output during processing
The goal is to create a zstd compressed output file with the RDF-XML
delimited by a comment tag that is compatible with the processing output
that narthex producesses.

The end-goal is to being able to process, index and publish this file
directly.

features:

- zstd compressed comment delimited file with all the RDF-XML records
- Full-view panel also shows the comment with identifiers, so that it
can be copy-and-pasted for testing
- the speed of processing should be minimally impacted to writing to a
single compressed file as compared to writing many single xml files
- sip-creator.properties should be expanded with the org_id

Change-Id: Id7910f98c6247f0652fb89bc3172bb5c603c5db0